+++ src/QualificationRunner.Core/Domain/Project.cs
@@ -6,6 +6,7 @@ namespace QualificationRunner.Core.Domain
    public class Project : IWithId
    {
       public string Id { get; set; }
+
       //Path of project. Either local path relative to qualification plan or a url
       public string Path { get; set; }
       public BuildingBlockRef[] BuildingBlocks { get; set; }
