+++ src/QualificationRunner.Core/Services/QualificationEngine.cs
@@ -15,31 +15,6 @@ using QualificationRunner.Core.RunOptions;
 
 namespace QualificationRunner.Core.Services
 {
-   public class QualificationRunResult : IReferencingProject
-   {
-      /// <summary>
-      ///    Path of the log file associated only with the current run
-      /// </summary>
-      public string LogFilePath { get; set; }
-
-      /// <summary>
-      ///    Path of the config file associated with the rin
-      /// </summary>
-      public string ConfigFile { get; set; }
-
-      /// <summary>
-      ///    Name of the mapping file created as a result of the qualification run
-      /// </summary>
-      public string MappingFile { get; set; }
-
-      /// <summary>
-      ///    Was the run successful
-      /// </summary>
-      public bool Success { get; set; }
-
-      public string Project { get; set; }
-   }
-
    public interface IQualificationEngine : IDisposable
    {
       Task<QualificationRunResult> Run(QualificationConfiguration qualificationConfiguration, QualificationRunOptions runOptions, ApplicationType application, CancellationToken cancellationToken);
